feat: Moved internal reconciliation db code to separate file to avoid oversized files
