+++ graphql/src/adapters/graphql/adapters.py
@@ -4,6 +4,8 @@ import dataclasses
 import strawberry
 from pydantic import BaseModel
 
+from src.adapters.storage.internal_reconciliation_rdbms import \
+    StorageAdapterInternalReconciliationRDBMS
 from src.adapters.storage.rdbms import StorageAdapterRDBMS
 from src.use_cases.adapter_interfaces.logger_provider import LoggerProviderInterface
 from src.use_cases.adapter_interfaces.word_generation import WordGenerationInterface
@@ -20,15 +22,19 @@ class AdaptersStorage(BaseModel):
 
     word_generation: WordGenerationInterface
     storage: StorageAdapterRDBMS
+    storage_internal_reconciliation: StorageAdapterInternalReconciliationRDBMS
     logger_provider: LoggerProviderInterface
 
     # Overriding constructor to give us type/name hints for Pydantic class.
     def __init__(self,
                  word_generation: WordGenerationInterface,
                  storage: StorageAdapterRDBMS,
+                 storage_internal_reconciliation: StorageAdapterInternalReconciliationRDBMS,
                  logger_provider: LoggerProviderInterface):
         # This call to __init__ will NOT automatically update when performing renames.
         super().__init__(
             word_generation=word_generation,
             storage=storage,
-            logger_provider=logger_provider)
+            storage_internal_reconciliation=storage_internal_reconciliation,
+            logger_provider=logger_provider
+        )
